Fix: SSR 수정
