+++ hooks/PrefetchCleanInfo.ts
@@ -1,0 +1,22 @@
+import axios from "axios";
+import { dehydrate, QueryClient } from "react-query";
+import { loadMyInfoAPI } from "../apis/user";
+import { queryKeys } from "../react-query/constants";
+
+export const PrefetchCleanInfo = async (context) => {
+  const cookie = context.req ? context.req.headers.cookie : '';
+  axios.defaults.headers.common.Cookie = '';
+  if (context.req && cookie) {
+    axios.defaults.headers.common.Cookie = cookie;
+  }
+  const queryClient = new QueryClient();
+
+  await queryClient.prefetchQuery(queryKeys.user, () => loadMyInfoAPI(cookie));
+
+  let cleanInfo = JSON.parse(JSON.stringify(dehydrate(queryClient)));
+  if (cleanInfo.queries[0].state.data && typeof cleanInfo.queries[0].state.data != 'undefined') {
+    cleanInfo.queries[0].state.data = cleanInfo.queries[0].state.data.data;
+  }
+
+  return cleanInfo;
+}
