+++ pages/introduce/index.tsx
@@ -5,10 +5,7 @@ import Map from '../../components/Map/map';
 import Link from 'next/link';
 import Head from 'next/head';
 import React from 'react';
-import axios from 'axios';
-import { dehydrate, QueryClient } from 'react-query';
-import { queryKeys } from '../../react-query/constants';
-import { loadMyInfoAPI } from '../../apis/user';
+import { PrefetchCleanInfo } from '../../hooks/PrefetchCleanInfo';
 
 const Introduce = () => {
   return (
@@ -104,19 +101,8 @@ const Introduce = () => {
 export default Introduce;
 
 export async function getServerSideProps(context) {
-  const cookie = context.req ? context.req.headers.cookie : '';
-  axios.defaults.headers.common.Cookie = '';
-  if (context.req && cookie) {
-    axios.defaults.headers.common.Cookie = cookie;
-  }
-  const queryClient = new QueryClient();
+  const cleanInfo = await PrefetchCleanInfo(context);
 
-  await queryClient.prefetchQuery(queryKeys.user, () => loadMyInfoAPI(cookie));
-
-  let cleanInfo = JSON.parse(JSON.stringify(dehydrate(queryClient)));
-  if (cleanInfo.queries[0].state.data && typeof cleanInfo.queries[0].state.data != 'undefined') {
-    cleanInfo.queries[0].state.data = cleanInfo.queries[0].state.data.data;
-  }
   return {
     props: {
       dehydratedState: cleanInfo,
